+++ Changelog.md
@@ -5,6 +5,7 @@ AppleALC Changelog
 - Added Desktop Intel Kaby Lake HD630 HDMI audio support (ig-platform-id 0x59120000)
 - Added Conexant CX20756 layout-id 13 by InsanelyDeepak
 - Added ALC282 layout-id 13 by InsanelyDeepak
+- Fixed separate layout/platform injection
 
 #### v1.1.2
 - Added ALC225 layout-id 28 for Dell Inspirion 7579 by ChalesYu